+++ beacon-chain/db/db_test.go
@@ -26,7 +26,7 @@ func setupDB(t testing.TB) *BeaconDB {
 	if err := os.RemoveAll(path); err != nil {
 		t.Fatalf("Failed to remove directory: %v", err)
 	}
-	db, err := NewDB(path)
+	db, err := NewDBDeprecated(path)
 	db.blocks = make(map[[32]byte]*ethpb.BeaconBlock)
 
 	if err != nil {
