+++ beacon-chain/rpc/validator_server_test.go
@@ -1014,7 +1014,7 @@ func BenchmarkAssignment(b *testing.B) {
 	b.StopTimer()
 	randPath, _ := rand.Int(rand.Reader, big.NewInt(1000000))
 	path := path.Join(testutil.TempDir(), fmt.Sprintf("/%d", randPath))
-	db, _ := db.NewDB(path)
+	db, _ := db.NewDBDeprecated(path)
 	defer db.Close()
 	os.RemoveAll(db.DatabasePath)
